Don't complain if a patch file doesn't exist for a particular version of python
I hope this isn't too controversial: I personally have this
`ASDF_PYTHON_PATCHES_DIRECTORY` globally set up on my machine, which is
super useful when I install/reinstall random versions of Python, but is
kind of annoying when I try to install versions of python that don't
actually need a patch: I have to create an empty file just to get it to
install.

With this change, I don't have to do that: instead I only have to create
`.patch` files for the versions of Python that need it.