Additional verbosity when using PYENV_DEBUG
I'm trying to figure out #2417 which hit me on 3.11.4 but not 3.11.3 which was
well after other reports. Having a way to debug seems useful either way.

Using `PYENV_DEBUG` may not seem like the best name but it works with the
underlying tools which avoids possible compatibility problems in the
future.